Initial RobotSuper commit. Adds a robot class that has a circular relationship with control loop.
This class provides the necessary structure for having a class that includes
a control loop with behaviors and more state information. The ciruclar dependency
may be more complicated than we want and so may not be the final solution.
Forward declarations and careful accounting of the pointers is necessary to avoid
build errors and seg faults.