implements the first version of a multiple inheritance robot class
Using the so called 'dreaded diamond' to implement a class that is derived from both an MjbotsRobotInterface and SimpleRobot example class which are each in turn virtually derived from kodlab::RobotInterface. This version requires building and passing in the robot to the control loop which upcasts the pointer to an MjbotsRobotInterface. Should not break legacy code