Provides abstract behavior class definition
Abstract behavior class is intended to provide a generic template for implementing robot behaviors. The function hooks are arranged to provide functionality during initialization and startup, as well as control loop updates and behavior transitions. Behaviors are intended to be stored by and transitioned between via a seperate behavior management object.