adds IMUData member to RobotBase
the IMUData is currently another shared pointer between the two classes, this can make construction a bit clunky, since we are not passing in the attitude class from ControlLoop. Instead now RobotBase and MJbotsHardwareInterface both take an IMUData shared_ptr as an optional argument in the constructor. The shared joints are more intuitive because they are passed in by the user to ControlLoop who then passes them along (by design) where attitude is not passed in by user. May want to change the interaction between the two or remove one of the constructors.