Add LcmSubscriber initialization method
LcmSubscriber inherits AbstractRealtimeObject, which requires starting of the thread on which it runs. This was previously done in the LcmSubsciber constructor, but this is not recommended because of constructor initialization order in derived classes, and generally indeterminate initialization orders (https://drake.mit.edu/styleguide/cppguide.html#Static_and_Global_Variables). This has been omved to a separate initialization function which must be called by the user.