Update constructors and example for imu_data
Updates the flow of robot and mjbots_interface constructors in MJBotsControlLoop.
New implementation changes the constructor of MJBotsControlLoop to have two optional arguments,
an IMU shared pointer and the world offset. The first if set to nullptr constructs and IMU, and
the second if not given does not update the worldoffset. This allows the user to prebuild the
IMUData class in the robot and choose how if worldoffset does anything. In MJBotsControlLoop,
the constructors are changed to use more delegation, and the LCM initialization is cleaned up
to match. Static casts were added to worldoffset to remove a narrowing warning. Finally,
SimpleRobot was updated to include examples of the state function calls.