Add timer and fix soft start construction parameter order in robot
This commit adds a realtime timer to remove the cycle_count that measure update
loops and fixes a construction bug where torque and duration were flipped.
In the case that the cycle time overflowed the intended frequency the
cycle count is no longer meaningful and gives the user little understanding of
the time past. A timer is agnostic to a misbehaving update cycle.