Inline LcmSubscriber implementation
Currently, the LcmSubscriber implementation is in the header file in which the class is defined.  If this file is included in multiple translation units, a linker error will occur due to multiple definitions of the member functions.  This will be fixed once the TODO in lcm_subscriber.h requesting that the implementation be moved to a cpp file is addressed.  However, this cannot happen until #51 is addressed.  This is a temporary fix in the meantime