Init commit for quad_3dof_limb example class
Limb example is based on kinematics of popular 12 DoF robotic
quadrupeds. In particular, a bespoke calculation of the limb kinematics
currently used in kodlab. May want to eventually update it to a more
standard approach, esp. checking the inverse kinematics perhaps DH.
This implementation could do with more comments expressing the
configuration description and perhaps enforcing a set of offsets for
clarity albeit loss of generality.