Fix error and saturation calculations and add overloads and setpoints
Add the setpoint members to the PID Class so that we don't need
to set them every loop, but give overloads that allow either.
Fixed the calculation for the integral and removed the numerical
derivative term. Fix the saturation logic using std::clamp and
changing the index which was bugged.