Fix function overrides, adjust usage, add typedef, comment cleanup
Fixed UpdateWithError which required zeroing out arguments, which
have to be const ref according to Eigen docs. So needed explicit copy.
The expectation was that the template would be used implicitly and also
that it would mostly be used in compositions as a member variable. These
two patterns don't work together nicely, so switching back to pushing
for explicit templating. Added a useful typedef for clarity and brevity
(RangeNS). Scalar gains now set all the gains per the templating
dimension, but we still error if the dimension doesn't match the state
input. Changed the name of the gains getter to be more intuitive.